theqoo-cosmetics-bot/requirements.txt
python-telegram-bot[job-queue]>=21.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
